customizable rate limiting; refactor utils
